♻️ Refactored hero section
